3.1-1の課題回答追加

+++ SampleProjects/3.1/MixiViewSample/MixiViewSample/MixiSecondViewController.m
@@ -40,6 +40,6 @@
 
 - (IBAction)pressButton:(UIButton *)sender
 {
-    NSLog(@"%d", sender.tag);
+    NSLog(@"%ld", (long)sender.tag);
 }
 @end
